cast tv_sec to long as we do everywhere else
Partially fixes #217 on OpenBSD 6.0

+++ ntpd/refclock_shm.c
@@ -519,8 +519,8 @@ shm_timer(
 	 * that interpretation is best left to higher levels.
 	 */
 	/* a_lastcode is seen as timecode with: ntpq -c cv [associd] */
-	c = snprintf(pp->a_lastcode, sizeof(pp->a_lastcode),
-		     "%ld.%09ld", shm_stat.tvt.tv_sec,	shm_stat.tvt.tv_nsec);
+	c = snprintf(pp->a_lastcode, sizeof(pp->a_lastcode), "%ld.%09ld",
+		     (long)shm_stat.tvt.tv_sec, shm_stat.tvt.tv_nsec);
 	pp->lencode = (c < (int)sizeof(pp->a_lastcode)) ? c : 0;
 
 	/* check 1: age control of local time stamp */
